+++ src/Entities/PaymentCardToken.php
@@ -11,28 +11,14 @@
 
 namespace Rebilly\Entities;
 
+use Rebilly\Entities\PaymentInstruments\PaymentCardPaymentInstrument;
 use Rebilly\Rest\Entity;
 
 /**
- * Class PaymentCardToken.
+ * @deprecated This class is considered deprecated in favour of PaymentToken
  */
-final class PaymentCardToken extends Entity
+class PaymentCardToken extends Entity
 {
-    /**
-     * @todo Rewrite ApiTest, which requires this method before deprecated methods.
-     *
-     * @param PaymentInstrument $value
-     *
-     * @return $this
-     */
-    public function setPaymentInstrument(PaymentInstrument $value)
-    {
-        $this->setAttribute('method', $value->name());
-        $this->setAttribute('paymentInstrument', $value->jsonSerialize());
-
-        return $this;
-    }
-
     /**
      * @deprecated The method is deprecated and will be removed in next version.
      * @see PaymentCardToken::setPaymentInstrument()
@@ -125,7 +111,7 @@ final class PaymentCardToken extends Entity
      */
     public function getFirstName()
     {
-        return $this->getAttribute('firstName');
+        return $this->getBillingAddressValue('firstName');
     }
 
     /**
@@ -136,7 +122,7 @@ final class PaymentCardToken extends Entity
      */
     public function setFirstName($value)
     {
-        return $this->setAttribute('firstName', $value);
+        return $this->setBillingAddressValue('firstName', $value);
     }
 
     /**
@@ -146,7 +132,7 @@ final class PaymentCardToken extends Entity
      */
     public function getLastName()
     {
-        return $this->getAttribute('lastName');
+        return $this->getBillingAddressValue('lastName');
     }
 
     /**
@@ -157,7 +143,7 @@ final class PaymentCardToken extends Entity
      */
     public function setLastName($value)
     {
-        return $this->setAttribute('lastName', $value);
+        return $this->setBillingAddressValue('lastName', $value);
     }
 
     /**
@@ -167,7 +153,7 @@ final class PaymentCardToken extends Entity
      */
     public function getAddress()
     {
-        return $this->getAttribute('address');
+        return $this->getBillingAddressValue('address');
     }
 
     /**
@@ -178,7 +164,7 @@ final class PaymentCardToken extends Entity
      */
     public function setAddress($value)
     {
-        return $this->setAttribute('address', $value);
+        return $this->setBillingAddressValue('address', $value);
     }
 
     /**
@@ -188,7 +174,7 @@ final class PaymentCardToken extends Entity
      */
     public function getAddress2()
     {
-        return $this->getAttribute('address2');
+        return $this->getBillingAddressValue('address2');
     }
 
     /**
@@ -199,7 +185,7 @@ final class PaymentCardToken extends Entity
      */
     public function setAddress2($value)
     {
-        return $this->setAttribute('address2', $value);
+        return $this->setBillingAddressValue('address2', $value);
     }
 
     /**
@@ -209,7 +195,7 @@ final class PaymentCardToken extends Entity
      */
     public function getCity()
     {
-        return $this->getAttribute('city');
+        return $this->getBillingAddressValue('city');
     }
 
     /**
@@ -220,7 +206,7 @@ final class PaymentCardToken extends Entity
      */
     public function setCity($value)
     {
-        return $this->setAttribute('city', $value);
+        return $this->setBillingAddressValue('city', $value);
     }
 
     /**
@@ -230,7 +216,7 @@ final class PaymentCardToken extends Entity
      */
     public function getRegion()
     {
-        return $this->getAttribute('region');
+        return $this->getBillingAddressValue('region');
     }
 
     /**
@@ -242,7 +228,7 @@ final class PaymentCardToken extends Entity
      */
     public function setRegion($value)
     {
-        return $this->setAttribute('region', $value);
+        return $this->setBillingAddressValue('region', $value);
     }
 
     /**
@@ -252,7 +238,7 @@ final class PaymentCardToken extends Entity
      */
     public function getCountry()
     {
-        return $this->getAttribute('country');
+        return $this->getBillingAddressValue('country');
     }
 
     /**
@@ -264,7 +250,7 @@ final class PaymentCardToken extends Entity
      */
     public function setCountry($value)
     {
-        return $this->setAttribute('country', $value);
+        return $this->setBillingAddressValue('country', $value);
     }
 
     /**
@@ -274,7 +260,7 @@ final class PaymentCardToken extends Entity
      */
     public function getPostalCode()
     {
-        return $this->getAttribute('postalCode');
+        return $this->getBillingAddressValue('postalCode');
     }
 
     /**
@@ -285,7 +271,7 @@ final class PaymentCardToken extends Entity
      */
     public function setPostalCode($value)
     {
-        return $this->setAttribute('postalCode', $value);
+        return $this->setBillingAddressValue('postalCode', $value);
     }
 
     /**
@@ -295,7 +281,7 @@ final class PaymentCardToken extends Entity
      */
     public function getPhoneNumber()
     {
-        return $this->getAttribute('phoneNumber');
+        return $this->getBillingAddressValue('phoneNumber');
     }
 
     /**
@@ -306,7 +292,15 @@ final class PaymentCardToken extends Entity
      */
     public function setPhoneNumber($value)
     {
-        return $this->setAttribute('phoneNumber', $value);
+        return $this->setBillingAddressValue('phoneNumber', $value);
+    }
+
+    /**
+     * @return Address
+     */
+    public function getBillingAddress()
+    {
+        return $this->getAttribute('billingAddress');
     }
 
     /**
@@ -317,6 +311,11 @@ final class PaymentCardToken extends Entity
     public function setBillingAddress($value)
     {
         return $this->setAttribute('billingAddress', $value);
+    }
+
+    public function createBillingAddress(array $value)
+    {
+        return new Address($value);
     }
 
     /**
@@ -356,31 +355,73 @@ final class PaymentCardToken extends Entity
     }
 
     /**
-     * @param $attribute
-     * @param $value
+     * @return PaymentMethodInstrument
+     */
+    public function getPaymentInstrument()
+    {
+        return $this->getAttribute('paymentInstrument');
+    }
+
+    /**
+     * @param PaymentInstrument $value
      *
      * @return $this
      */
-    private function setDefaultPaymentInstrumentValue($attribute, $value)
+    public function setPaymentInstrument(PaymentInstrument $value)
     {
-        $data = (array) $this->getAttribute('paymentInstrument');
-        $data[$attribute] = $value;
-
-        $this->setAttribute('method', PaymentMethod::METHOD_PAYMENT_CARD);
-        $this->setAttribute('paymentInstrument', $data);
+        $this->setAttribute('method', $value->name());
+        $this->setAttribute('paymentInstrument', $value);
 
         return $this;
     }
 
     /**
-     * @param $attribute
+     * @param array $data
      *
-     * @return mixed|null
+     * @return PaymentInstrument
      */
-    private function getDefaultPaymentInstrumentValue($attribute)
+    public function createPaymentInstrument(array $data)
     {
-        $data = (array) $this->getAttribute('paymentInstrument');
+        return PaymentInstrument::createFromData($data);
+    }
 
-        return $data[$attribute] ?? null;
+    private function setDefaultPaymentInstrumentValue(string $attribute, $value)
+    {
+        if (!$this->getAttribute('paymentInstrument')) {
+            $this->setPaymentInstrument(new PaymentCardPaymentInstrument());
+        }
+
+        $this->getPaymentInstrument()->setAttribute($attribute, $value);
+
+        return $this;
+    }
+
+    private function getDefaultPaymentInstrumentValue(string $attribute)
+    {
+        if (!$this->getAttribute('paymentInstrument')) {
+            $this->setPaymentInstrument(new PaymentCardPaymentInstrument());
+        }
+
+        return $this->getPaymentInstrument()->getAttribute($attribute);
+    }
+
+    private function setBillingAddressValue(string $attribute, $value)
+    {
+        if (!$this->getAttribute('billingAddress')) {
+            $this->setBillingAddress([]);
+        }
+
+        $this->getBillingAddress()->setAttribute($attribute, $value);
+
+        return $this;
+    }
+
+    private function getBillingAddressValue(string $attribute)
+    {
+        if (!$this->getAttribute('billingAddress')) {
+            $this->setBillingAddress([]);
+        }
+
+        return $this->getBillingAddress()->getAttribute($attribute);
     }
 }